refactor: exit to utils
